pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aa-metenox"
dynamic = ["version", "description"]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
authors = [
    { name = "T'rahk Rokym", email = "trahk.rokym@gmail.com" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth>=4.6.0",
    "aa-moonmining>=1.11",
    "allianceauth-app-utils>=1.25.0",
    "django-datatables-view>=1.20.0",
    "dhooks-lite>=1.1.0",
    "responses>=0.25",
]


[project.urls]
Homepage = "https://gitlab.com/r0kym/aa-metenox"
Source = "https://gitlab.com/r0kym/aa-metenox"
Changelog = "https://gitlab.com/r0kym/aa-metenox/-/blob/main/CHANGELOG.md"
Tracker = "https://gitlab.com/r0kym/aa-metenox/-/issues"

[tool.flit.module]
name = "metenox"

[tool.isort]
profile = "black"
multi_line_output = 3
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_allianceauth = ["allianceauth", "app_utils"]
known_django = ["django", "django_webtest", "esi", "eveuniverse"]

[tool.pylint.'MASTER']
ignore-patterns = ["__init__.py", "auth_hooks.py", "apps.py", "admin.py"]
ignore-paths = ["^.*/tests/.*$", "^.*/migrations/.*$"]

[tool.pylint.'BASIC']
good-names = ["i", "j", "k", "ex"]

[tool.pylint.'FORMAT']
max-line-length = 120

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "cyclic-import",
    "fixme",
    "import-outside-toplevel",
    "imported-auth-user",
    "too-few-public-methods",
]
